Extract plugin-api chain module
Move the blockchain query, RLP conversion and block-view contracts into a new
ethereum/core/plugin-api-chain module, published as besu-plugin-api-chain.
Packages are untouched and plugin-api re-exports the module, so the surface
reachable through besu-plugin-api is unchanged.

BlockchainServiceImpl and RlpConverterServiceImpl move to
org.hyperledger.besu.ethereum.chain.pluginadapter, as permissioning and
security did.

Corrects three javadoc errors on the moved contracts: the BlockchainService
class comment described one of its eighteen methods and omitted the writes,
setSafeBlock's @param named the finalized block, and RlpConverterService had
a misspelling.
