+++ app/src/main/java/org/hyperledger/besu/services/BesuPluginServiceRegistrar.java
@@ -18,6 +18,7 @@ import org.hyperledger.besu.Runner;
 import org.hyperledger.besu.controller.BesuController;
 import org.hyperledger.besu.ethereum.api.pluginadapter.HealthCheckServiceImpl;
 import org.hyperledger.besu.ethereum.api.query.BlockchainQueries;
+import org.hyperledger.besu.ethereum.chain.pluginadapter.RlpConverterServiceImpl;
 import org.hyperledger.besu.ethereum.core.MiningConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.pluginadapter.TransactionPoolServiceImpl;
 import org.hyperledger.besu.plugin.services.BesuEvents;
