+++ ethereum/core/plugin-api-chain/build.gradle
@@ -1,0 +1,56 @@
+/*
+ * Copyright contributors to Besu.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'java-library'
+
+// override the path-derived name (besu-ethereum-core-plugin-api-chain) with the plugin API family naming
+def chainApiArtifactId = 'besu-plugin-api-chain'
+
+jar {
+  archiveBaseName = chainApiArtifactId
+  manifest {
+    attributes(
+      'Specification-Title': archiveBaseName,
+      'Specification-Version': project.version,
+      'Implementation-Title': archiveBaseName,
+      'Implementation-Version': calculateVersion(),
+      'Commit-Hash': getGitCommitDetails(40).hash,
+      'Automatic-Module-Name': 'org.hyperledger.besu.plugin.api.chain'
+      )
+  }
+}
+
+dependencies {
+  api project(':plugin-api:core')
+  // Hash, Wei, Transaction, HardforkId, Address and LogTopic appear in this module's own
+  // signatures, so the dependency is declared here even though plugin-api:core re-exports it
+  api project(':datatypes')
+  // Bytes is in RlpConverterService and LogWithMetadata, UInt256 in PropagatedBlockContext
+  api 'io.consensys.tuweni:tuweni-bytes'
+  api 'io.consensys.tuweni:tuweni-units'
+}
+
+publishing {
+  publications {
+    mavenJava(MavenPublication) {
+      groupId = 'org.hyperledger.besu'
+      artifactId = chainApiArtifactId
+      pom {
+        name = 'Besu Plugin API Chain'
+        description = 'Blockchain contracts of the Besu plugin API: the blockchain service for chain queries and chain markers, the RLP conversion service, and the block, log and bad-block data views'
+      }
+    }
+  }
+}
