+++ ethereum/core/plugin-api-chain/src/main/java/org/hyperledger/besu/plugin/data/LogWithMetadata.java
