+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/chain/pluginadapter/RlpConverterServiceImplTest.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.services;
+package org.hyperledger.besu.ethereum.chain.pluginadapter;
 
 import static org.assertj.core.api.Assertions.assertThat;
