+++ platform/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     api project(':ethereum:api:plugin-api')
     api project(':ethereum:blockcreation')
     api project(':ethereum:core')
+    api project(':ethereum:core:plugin-api-chain')
     api project(':ethereum:eth')
     api project(':ethereum:eth:plugin-api-sync')
     api project(':ethereum:eth:plugin-api-txpool')
