+++ plugin-api/build.gradle
@@ -37,6 +37,7 @@ dependencies {
   api project(':ethereum:eth:plugin-api-txpool')
   api project(':ethereum:eth:plugin-api-sync')
   api project(':ethereum:api:plugin-api')
+  api project(':ethereum:core:plugin-api-chain')
   api project(':datatypes')
   api 'org.apache.commons:commons-lang3'
   api 'io.consensys.tuweni:tuweni-bytes'
@@ -72,7 +73,8 @@ tasks.register('checkAPICompatibility', me.champeau.gradle.japicmp.JapicmpTask) 
     project(':ethereum:p2p:plugin-api').tasks.named('jar'),
     project(':ethereum:eth:plugin-api-txpool').tasks.named('jar'),
     project(':ethereum:eth:plugin-api-sync').tasks.named('jar'),
-    project(':ethereum:api:plugin-api').tasks.named('jar'))
+    project(':ethereum:api:plugin-api').tasks.named('jar'),
+    project(':ethereum:core:plugin-api-chain').tasks.named('jar'))
   newClasspath.from(sourceSets.main.runtimeClasspath)
 
   packageIncludes = [
