+++ settings.gradle
@@ -43,6 +43,7 @@ include 'ethereum:api'
 include 'ethereum:api:plugin-api'
 include 'ethereum:blockcreation'
 include 'ethereum:core'
+include 'ethereum:core:plugin-api-chain'
 include 'ethereum:eth'
 include 'ethereum:eth:plugin-api-sync'
 include 'ethereum:eth:plugin-api-txpool'
